data-processor/requirements.txt
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.3
scikit-learn==0.24.2